+++ CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [CalVer](https://calver.org/).
 
+## [UNRELEASED] - 2025-18-3
+
+### Added
+
+- A new config `editorAlwaysActivated`. If set to `false` the editor controls will be hidden until the user sign in.
+
+### Fixed
+
+- Better handling of vector layers in `activeLayers`. If `activeLayers` only had vector layers, they wouldn't be activated.
+
 ## [2025.3.1] - 2025-10-3
 
 ### Added